Création d'une commande abstraite BuildableCmd pour factoriser la construction d'un site
